update test cases and change the index & response
