+++ .eslintrc.json
@@ -10,22 +10,12 @@
     "parser": "@typescript-eslint/parser",
     "parserOptions": {
         "ecmaVersion": "latest",
-        "sourceType": "module",
-        "project": ["tsconfig.json"] 
+        "sourceType": "module"
     },
     "plugins": [
         "@typescript-eslint"
     ],
     "rules": {
-        "@typescript-eslint/promise-function-async": [
-            "error",
-            {
-              "allowedPromiseNames": ["Thenable"],
-              "checkArrowFunctions": true,
-              "checkFunctionDeclarations": true,
-              "checkFunctionExpressions": true,
-              "checkMethodDeclarations": true
-            }
-        ]
+
     }
 }
