+++ test/lib/saml20.responseSignedMessage.spec.ts
@@ -1,4 +1,4 @@
-import saml from '../../lib/index';
+import { validate } from '../../lib/response';
 import { expect } from 'chai';
 import fs from 'fs';
 
@@ -13,138 +13,81 @@ const audience = 'http://sp.example.com/demo1/metadata.php';
 const inResponseTo = 'ONELOGIN_4fee3b046395c4e751011e97f8900b5273d56685';
 
 describe('saml20.responseSignedMessage', function () {
-  it('Should validate saml 2.0 token using thumbprint', function (done) {
-    saml.validateInternal(
-      validResponse,
-      {
-        publicKey: certificate,
-        thumbprint: thumbprint,
-        bypassExpiration: true,
-        inResponseTo: inResponseTo,
-      },
-      function (err, profile) {
-        expect(profile.claims).to.be.ok;
-        expect(err).to.not.be.ok;
+  it('Should validate saml 2.0 token using thumbprint', async function () {
+    const response = await validate(validResponse, {
+      publicKey: certificate,
+      thumbprint: thumbprint,
+      bypassExpiration: true,
+      inResponseTo: inResponseTo,
+    });
 
-        expect(issuerName).to.equal(profile.issuer);
-        expect('_ce3d2948b4cf20146dee0a0b3dd6f69b6cf86f62d7').to.equal(
-          profile.claims['http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier']
-        );
-
-        done();
-      }
+    expect(issuerName).to.equal(response.issuer);
+    expect('_ce3d2948b4cf20146dee0a0b3dd6f69b6cf86f62d7').to.equal(
+      response.claims['http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier']
     );
   });
 
-  it('Should validate saml 2.0 token using certificate', function (done) {
-    saml.validateInternal(
-      validResponse,
-      {
-        publicKey: certificate,
-        bypassExpiration: true,
-        inResponseTo: inResponseTo,
-      },
-      function (err, profile) {
-        expect(profile.claims).to.be.ok;
-        expect(err).to.not.be.ok;
+  it('Should validate saml 2.0 token using certificate', async function () {
+    const response = await validate(validResponse, {
+      publicKey: certificate,
+      bypassExpiration: true,
+      inResponseTo: inResponseTo,
+    });
 
-        expect(issuerName).to.equal(profile.issuer);
-        expect('_ce3d2948b4cf20146dee0a0b3dd6f69b6cf86f62d7').to.equal(
-          profile.claims['http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier']
-        );
-
-        done();
-      }
+    expect(issuerName).to.equal(response.issuer);
+    expect('_ce3d2948b4cf20146dee0a0b3dd6f69b6cf86f62d7').to.equal(
+      response.claims['http://schemas.xmlsoap.org/ws/2005/05/identity/claims/nameidentifier']
     );
   });
 
-  it('Should validate saml 2.0 token and check audience', function (done) {
-    saml.validateInternal(
-      validResponse,
-      {
-        publicKey: certificate,
-        audience: audience,
-        bypassExpiration: true,
-        inResponseTo: inResponseTo,
-      },
-      function (err, profile) {
-        expect(profile.claims).to.be.ok;
-        expect(err).to.not.be.ok;
-
-        expect(issuerName).to.equal(profile.issuer);
-        done();
-      }
-    );
+  it('Should validate saml 2.0 token and check audience', async function () {
+    const response = await validate(validResponse, {
+      publicKey: certificate,
+      audience: audience,
+      bypassExpiration: true,
+      inResponseTo: inResponseTo,
+    });
+    expect(issuerName).to.equal(response.issuer);
   });
 
-  it('Should fail with invalid audience', function (done) {
-    saml.validateInternal(
-      validResponse,
-      {
+  it('Should fail with invalid audience', async function () {
+    try {
+      await validate(validResponse, {
         publicKey: certificate,
         audience: 'http://any-other-audience.com/',
         bypassExpiration: true,
         inResponseTo: inResponseTo,
-      },
-      function (err, profile) {
-        expect(profile).to.not.be.ok;
-        expect(err).to.be.ok;
-        try {
-          if (err) {
-            expect(err.message).to.equal('Invalid assertion.');
-          }
-          done();
-        } catch (error) {
-          done();
-        }
-      }
-    );
+      });
+    } catch (error) {
+      const result = (error as Error).message;
+      expect(result).to.be.equal('Invalid audience.');
+    }
   });
 
-  it('Should fail with invalid assertion', function (done) {
-    saml.validateInternal(
-      'invalid-assertion',
-      {
+  it('Should fail with invalid assertion', async function () {
+    try {
+      await validate('invalid-assertion', {
         publicKey: certificate,
         bypassExpiration: true,
         inResponseTo: inResponseTo,
-      },
-      function (err, profile) {
-        expect(profile).to.not.be.ok;
-        expect(err).to.be.ok;
-        try {
-          if (err) {
-            expect(err.message).to.equal('Invalid assertion.');
-          }
-          done();
-        } catch (error) {
-          done();
-        }
-      }
-    );
+      });
+    } catch (error) {
+      const result = (error as Error).message;
+      expect(result).to.be.equal('Invalid assertion.');
+    }
   });
 
-  it('Should fail with invalid inResponseTo', function (done) {
-    saml.validateInternal(
-      validResponse,
-      {
+  it('Should fail with invalid inResponseTo', async function () {
+    try {
+      await validate(validResponse, {
         publicKey: certificate,
         audience: audience,
         bypassExpiration: true,
         inResponseTo: 'not-the-right-response-to',
-      },
-      function (err, profile) {
-        expect(profile).to.not.be.ok;
-        expect(err).to.be.ok;
-        try {
-          if (err) {
-            expect(err.message).to.equal('Invalid InResponseTo.');
-          }
-          done();
-        } catch (error) {
-          done();
-        }
-      }
-    );
+      });
+    } catch (error) {
+      const result = (error as Error).message;
+      expect(result).to.be.equal('Invalid InResponseTo.');
+    }
   });
 });
